Close to releaseable form with xxHash
